+++ example/lib/full_screen_ads.dart
@@ -114,7 +114,7 @@ class _FullScreenAdsState extends State<FullScreenAds> {
       RewardedAd rewardedAd = (wrapper as RewardedWrapper).ad;
       request = () => rewardedAd.request();
       show = () =>
-          rewardedAd.showWithSSR(serverSideRewarding: {"option1": "value1"});
+          rewardedAd.showWithSSR(serverSideRewardingOptions: {"option1": "value1"});
       available = rewardedAd.isAvailable
           .asStream()
           .concatWith([rewardedAd.availabilityStream]);
@@ -124,7 +124,7 @@ class _FullScreenAdsState extends State<FullScreenAds> {
     } else {
       InterstitialAd interstitialAd = (wrapper as InterstitialWrapper).ad;
       request = () => interstitialAd.request();
-      show = () => interstitialAd.show();
+      show = () => interstitialAd.showWithSSR(serverSideRewardingOptions: {"option1":"value1"});
       available = interstitialAd.isAvailable
           .asStream()
           .concatWith([interstitialAd.availabilityStream]);